+++ src/main/java/survivalblock/atmosphere/atmospheric_api/mixin/item/spyglass/client/InGameHudAccessor.java
@@ -8,6 +8,7 @@ import org.spongepowered.asm.mixin.gen.Accessor;
 @Mixin(Gui.class)
 public interface InGameHudAccessor {
 
+    @SuppressWarnings("unused")
     @Accessor("SPYGLASS_SCOPE_LOCATION")
     static ResourceLocation atmospheric_api$getSpyglassOverlay() {
         throw new UnsupportedOperationException();