+++ src/main/java/survivalblock/atmosphere/atmospheric_api/not_mixin/resource/client/AdditionalSplashTextRegistry.java
@@ -1,0 +1,19 @@
+package survivalblock.atmosphere.atmospheric_api.not_mixin.resource.client;
+
+import net.fabricmc.api.EnvType;
+import net.fabricmc.api.Environment;
+
+@Environment(EnvType.CLIENT)
+public final class AdditionalSplashTextRegistry {
+
+    private AdditionalSplashTextRegistry() {
+    }
+
+    public static void register(String string) {
+
+    }
+
+    public static void register(String string, int priority) {
+
+    }
+}